requirements.txt
# Grokgates v1 Dependencies

# Core async HTTP client
httpx==0.27.0

# Redis for shared state
redis==5.0.1

# Rich terminal UI
rich==13.7.0

# CLI interface
click==8.1.7

# Web server dependencies
flask==3.0.0
flask-socketio==5.3.5
flask-cors==4.0.0
python-socketio==5.10.0

# Vector database for agent memory
chromadb==0.4.22

# Env loader
python-dotenv==1.0.1

# Python version requirement: 3.11+